requirements.txt
backcall==0.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
cycler==0.11.0
debugpy==1.7.0
decorator==5.1.1
entrypoints==0.4
fonttools==4.38.0
idna==3.7
ipykernel==6.16.2
ipython==7.34.0
jedi==0.19.1
jupyter_client==7.4.9
jupyter_core==4.12.0
kiwisolver==1.4.5
matplotlib==3.5.3
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==24.0
pandas==1.3.5
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==25.1.2
requests==2.31.0
six==1.16.0
torch==1.13.1
torchsummary==1.5.1
torchvision==0.14.1
tornado==6.2
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==2.0.7
wcwidth==0.2.13
